Big overhaul of internal representation of irrelevant subterms.
Ignore-this: 2c41af95e243dd3969f1b6a4a7b29940
DontCare is not longer inserted everywhere.
Irrelevant subterms are distinguished by the Relevance field in Arg.

DontCare has been rededicated to an internal represent of the
irrelevance axiom: .DontCare : .A -> A
It is used to guard the content of irrelevant projections.
The previous solution, to use a user-postulated builtin irrelevance
axiom, does not work since we cannot supply the type A; it is
gone due to the erasure of parameters in projections.
The dummy (sort Prop) solution failed in the presence of 'with'.
This was raised in issue 596 which is now fixed.

Since Arg is now used to distinguish irrelevant subterms, it has
been split into Arg and Dom.  Dom is used in Pi-types, contexts and
telescopes where the relevance flag does not apply to the subterm
(function domain) itself, but is an information about the function
type variance.  Arg is used whenever the relevance flag applies to
the term under Arg itself.

Since Arg has been used for many things, there are cases where both
Dom and Arg can be used.  In case of doubt, use Arg.

darcs-hash:20120403141629-cc7ee-da18145cd841f567d7a94f7917d869b7b14c7411