Enhanced pruning by makeing hasBadRigid more precise.
Ignore-this: 762ab7e24dcd0e317891d9546471b60d
Any argument to a metavariable which has a forbidden variable in a non-eliminateable position (e.g. as an argument to a type constructor) can be pruned.
Before it was just a forbidden variable in head position. (Which is a special case of a non-eliminateable position.)

darcs-hash:20120504114320-5187b-7a343772964df9b3702aaa8be4f587d646b911d1