Termination checker now correctly rejects infinite objects in a mu-nu type, if defined via copatterns.
Ignore-this: 382329378a3d7a4970167ba861faf218
The termination checker now only counts coconstructors as guarding if they
belong to the record type that is mutually recursive with the target type
of the function it is currently checking.

For example, this code, taken from Nisse and Thorstens PAR 2010 note,
is no longer accepted (see test/fail/ColistMutual.agda):

  mutual

    data CoList (A : Set) : Set where
      []  : CoList A
      _∷_ : (x : A)(xs : ∞CoList A) → CoList A

    record ∞CoList (A : Set) : Set where
      coinductive
      field out : CoList A

  open ∞CoList

  data Tree : Set where
    node : CoList Tree → Tree

  mutual

    bad : Tree
    bad = node (node [] ∷ bads)

    bads : ∞CoList Tree
    out bads = bad ∷ bads

In this case, the coconstructor for ∞CoList does not guard function
bad, since it has target Tree which is not mutually recursive with
∞CoList.

To achieve this with minimal changes to the termination checker, it
now checks each function in a mutual block individually.

Caveat: this leads to a 15% loss in performance on the standard library.
The performance loss is only transitional during the experimentation stage.
If the new termination technique solves the problems of the previous
termination checker, we can implement a new call graph generation which
does not have the current waste of performance.

darcs-hash:20120619145205-5187b-ee0973d3ef651a7b4ec67c6a9e7eeac9e27a77ec