New error message for issue 543, probably caused by syntactic equality check.
Ignore-this: 6c6535bcc2bbe293a949ee52ce68818e
The new error message seems to make more sense.  Basically, the subst
expression which is now underlined has type

  .(true) == .(false)

and when comparing this to the goal

  true == false

Agda realizes that .(true) is not equal to true.  The old message

  .(true) != .(true)

was owed to the fact that DontCare things were never considered equal,
even if they are identical, but this looks puzzling at first.

darcs-hash:20121012173342-5187b-d3e96217f52870dd77013d32d4a3ba61a4fcfae8